warnings was removed
git-svn-id: svn://localhost/trunk@1013 3bdad9f7-c964-442a-868d-553b97b55ca0

+++ surfit/src/surfit/cmofs.cpp
@@ -126,7 +126,7 @@ void surfit() {
 
 	};
 
-	int f_size = functionals->size();
+	size_t f_size = functionals->size();
 
 	for (i = 0; i < f_size; i++) {
 		functional * f = (*functionals)[i];
